Fix benchmarks for polarfire (#81)
* Add WASM support for bpf_rand_int, update matmul test

* Add riscv compilation target for wasmtime

* Fix bugs in tests

* Recompile all tests

* Fix bugs with NccScore test

* Improve populate_map_random

* Improve how test results are outputted

* Fix bpf NccScore test and use bpf_iter

* Update tests for iter changes, uncomment fixed Aberr test

+++ tests/kalman/program.bpf.c
@@ -4,7 +4,7 @@ int main() {
     void *in_map = MAP_BY_FD(3), *out_map = MAP_BY_FD(4), *result;
     struct bpf_iter_num it;
     float ins[7], preds[7];
-    int* i;
+    long long* i;
 
     // Read in position and attitude
     bpf_iter_num_new(&it, 0, 7);
